% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_styling.R
\name{add_beis_body_style}
\alias{add_beis_body_style}
\title{Function for styling the main body of an excel table}
\usage{
add_beis_body_style(
  wb,
  sheet = 1,
  rows = 2:6,
  cols = 1:3,
  showGridLines = FALSE,
  ...
)
}
\arguments{
\item{wb}{the workbook to add the stylying to as generarated from the openxlsx package}

\item{sheet}{the sheet to apply the styling to, can be a number or a string}

\item{rows}{the row to apply the syling to}

\item{cols}{the columns to apply the syling to}

\item{showGridLines}{Whether to show gridline in the worksheet. Default is FALSE.}

\item{...}{additional parameters you might want to add from openxlsx::create_style()}
}
\value{
an updated workbook object
}
\description{
Styled based on recommended defaults by CAGG Team. Will remove gridlines for sheet applied by default
}
\details{
#' Uses the default beis blue "#003366"
}
\examples{
#create some dummy data
trade <-
 data.frame(
   Country = c("France", "Germany", "Italy", "Spain", "UK", "US"),
    Imports = c(100000L, 150000L, 270000L, 123000L, 790000L, 5678900L),
    Exports = c(230000L, 120000L, 870000L, 738000L, 654000L, 65678000L)
  )

# create the workbook and write in the worksheet with the data
wb <- openxlsx::createWorkbook()
openxlsx::addWorksheet(wb, "Trade")
openxlsx::writeData(wb, "Trade", trade)

add_beis_header_style(wb, header_classic = FALSE)
add_beis_body_style(wb, textDecoration = "italic")

##NOT RUN
#saveWorkbook(wb = wb, file =  "P:/wb_test.xlsx", overwrite = TRUE)
}
