% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_styling.R
\name{add_beis_header_style}
\alias{add_beis_header_style}
\title{Add a header in a 'beis' style format to your workbook}
\usage{
add_beis_header_style(
  wb,
  sheet = 1,
  rows = 1,
  cols = 1:3,
  header_classic = TRUE,
  halign = "right",
  ...
)
}
\arguments{
\item{wb}{the workbook to add the stylying to as generarated from the openxlsx package}

\item{sheet}{the sheet to apply the styling to}

\item{rows}{the row to apply the syling to}

\item{cols}{the columns to apply the syling to}

\item{header_classic}{whether to use a classic beis header with just the bottom border or
a pronounced header with blue background and white text}

\item{halign}{rigght align by default}

\item{...}{additional parameters you might want to add from openxlsx::create_style()}
}
\value{
an updated workbook object
}
\description{
There are two header styles available. One with a bottom border and blue text and another
with a filled blue background and bold white text.
}
\details{
Uses the default beis blue "#003366"
}
\examples{
#create some dummy data
trade <-
 data.frame(
   Country = c("France", "Germany", "Italy", "Spain", "UK", "US"),
    Imports = c(100000L, 150000L, 270000L, 123000L, 790000L, 5678900L),
    Exports = c(230000L, 120000L, 870000L, 738000L, 654000L, 65678000L)
  )

# create the workbook and write in the worksheet with the data
wb <- openxlsx::createWorkbook()
openxlsx::addWorksheet(wb, "Trade")
openxlsx::writeData(wb, "Trade", trade)


add_beis_header_style(wb, header_classic = FALSE)
add_beis_body_style(wb, textDecoration = "italic")

##NOT RUN
#saveWorkbook(wb = wb, file =  "P:/wb_test.xlsx", overwrite = TRUE)
}
