% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beis_colours.R
\name{beis_colours}
\alias{beis_colours}
\title{Get BEIS colour palette.}
\usage{
beis_colours(n = 15, distinct = FALSE)
}
\arguments{
\item{n}{The number of colours to return, max is 15 for non-distinct, 5 for distinct.}

\item{distinct}{Whether to return only colours which are more easily distinguishable or all colours}
}
\value{
Returns a character vector of the hex codes for BEIS colours

BEIS colours:
Dark blue = #003366
Bright blue = #0099cc
Grey = #a8b8be
cobalt blue = #0047ab
cyan = #689d9c
Magenta (pink) = #ee3d8f - NOT USED
Green = #9ccd63  - Not USED
Dark moss green = #4a5d23
Teal = #006666
Platinum = #c2baa9 - Noy used
Medium dark blue = #3d4556

Non-BEIS colours also included:
Light blue = #c5e8ff - Not used
Root beer = #290e05
Pistachio green = #93c572
Darker grey = #505050
Red = #ec3337
Darker red = #9a0e11
Gold = #fbb034
White = #FFFFFF - Not Used
Black = #000000

The ONS recommend three helpful rules when choosing colours: (1) Get it right in black and white, (2) The safest hue is blue and (3) Red and green should never be seen (together). For more on choosing colour palettes, see ONS's introduction on 'Using colours', available on their website
}
\description{
Retrieves a BEIS-like colour palette as a character vector.
}
\examples{
beis_colours(n = 3, distinct = TRUE)
}
