% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beis_css.R
\name{beis_css}
\alias{beis_css}
\title{Copy & call BEIS css from working directory.}
\usage{
beis_css(copy_to = getwd())
}
\arguments{
\item{copy_to}{The working directory to copy the css to, default current working directory}
}
\value{
Returns a copy of the png image to the directory AND calls the image filepath for use
}
\description{
Calls \code{get_beis_css()} with default and then uses the resulting filepath.
}
\examples{
beis_css(copy_to = "my_folder")

}
\seealso{
get_beis_css
}
