% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beis_logo.R
\name{beis_logo}
\alias{beis_logo}
\title{Copy & call BEIS logo from working directory.}
\usage{
beis_logo(copy_to = getwd(), white_type = FALSE)
}
\arguments{
\item{copy_to}{The working directory to copy the logo to, default current working directory}

\item{white_type}{Whether or not to copy the white typeface version, default FALSE}
}
\value{
Returns a copy of the png image to the directory AND calls the image filepath for use
}
\description{
Calls \code{get_beis_logo()} with default and then uses the resulting filepath.
}
\examples{
beis_logo(copy_to = "my_folder", white_type = FALSE)

}
\seealso{
get_beis_logo
}
