% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beis_css.R
\name{get_beis_css}
\alias{get_beis_css}
\title{Includes BEIS css stylesheet.}
\usage{
get_beis_css(copy_to = getwd())
}
\arguments{
\item{copy_to}{The working directory to copy the css file to, default current working directory}
}
\value{
css file to style elements of markdown and dashboards.
}
\description{
Copies BEIS-styling css stylesheet to specified directory.
}
\examples{
get_beis_css(copy_to = getwd())
}
