% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beis_logo.R
\name{get_beis_logo}
\alias{get_beis_logo}
\title{Copy BEIS logo to working directory.}
\usage{
get_beis_logo(copy_to = getwd(), white_type = FALSE)
}
\arguments{
\item{copy_to}{The working directory to copy the logo to, default current working directory}

\item{white_type}{Whether or not to copy the white typeface version, default FALSE}
}
\value{
Returns a copy of the png image to the directory
}
\description{
Copies the standard or white typeface version of the BEIS logo to your current working directory.
}
\examples{
get_beis_logo(copy_to = "my_folder", white_type = FALSE)
}
