% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_beis.R
\name{theme_beis}
\alias{theme_beis}
\title{Style a ggplot with a BEIS theme.}
\usage{
theme_beis()
}
\value{
Returns a styled ggplot when used as the theme
}
\description{
Uses the ggplot2::theme() function to create a BEIS style for ggplots.
}
\examples{
ggplot2::ggplot(mtcars) +
ggplot2::geom_point(ggplot2::aes(x = wt, y = mpg, colour = gear)) +
theme_beis()
}
