% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_beis_dygraph.R
\name{theme_beis_dygraph}
\alias{theme_beis_dygraph}
\title{Style a dygraph with a BEIS theme.}
\usage{
theme_beis_dygraph(dygraph, ...)
}
\arguments{
\item{dygraph}{The dygraph object}

\item{...}{Any additional arguments to pass to dygraph or beis_css or beis_colours()}
}
\value{
Returns a styled dygraph when piped to
}
\description{
Uses the dygraphs package and calls beis_css() and beis_colours() to style
}
\examples{
dygraphs::dygraph(cbind(mdeaths, fdeaths), main = "title") \%>\%
theme_beis_dygraph()
}
