% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_beis.R
\name{theme_beis}
\alias{theme_beis}
\title{Style a ggplot with a BEIS theme.}
\usage{
theme_beis(default_colours = TRUE, distinct_colours = FALSE, ...)
}
\arguments{
\item{default_colours}{whether to use the default colour scheme for BEIS or the alternative}

\item{distinct_colours}{Whether to use Dsitinct colours - maximum of 5}

\item{...}{Any additional arguments to pass to ggtheme}
}
\value{
Returns a styled ggplot when used as the theme
}
\description{
Uses the ggplot2::theme() function to create a BEIS style for ggplots.
}
\examples{
ggplot2::ggplot(mtcars) +
ggplot2::geom_point(ggplot2::aes(x = wt, y = mpg, colour = as.factor(gear))) +
theme_beis()
}
