% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_beis_dygraph.R
\name{theme_beis_dygraph}
\alias{theme_beis_dygraph}
\title{Style a dygraph with a BEIS theme.}
\usage{
theme_beis_dygraph(dygraph, copy_to = getwd())
}
\arguments{
\item{dygraph}{The dygraph object}

\item{copy_to}{The working directory to copy the logo to, default current working directory}

\item{...}{Any additional arguments to pass to dygraph or beis_css or beis_colours()}
}
\value{
Returns a styled dygraph when piped to
}
\description{
Uses the dygraphs package and calls beis_css() and beis_colours() to style
}
\examples{
d <- dygraphs::dygraph(cbind(mdeaths, fdeaths), main = "title")
p <- theme_beis_dygraph(d, copy_to = "P:/")

}
\seealso{
beis_css, beis_colours
}
